requirements.txt
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
pyzbar==0.1.9
